fix(gemini): cast sizeBytes to int in GeminiFile
Google's API returns sizeBytes as a string (e.g., "5806") but the
GeminiFile constructor expects an int. Added explicit type cast to
convert the string to integer.

Fixes TypeError: Argument #4 ($sizeBytes) must be of type int,
string given

+++ src/Providers/Gemini/ValueObjects/GeminiFile.php
@@ -33,7 +33,7 @@ readonly class GeminiFile
             name: data_get($response, 'name', ''),
             displayName: data_get($response, 'displayName', ''),
             mimeType: data_get($response, 'mimeType', ''),
-            sizeBytes: data_get($response, 'sizeBytes', 0),
+            sizeBytes: (int) data_get($response, 'sizeBytes', 0),
             createTime: Carbon::parse(data_get($response, 'createTime')),
             updateTime: Carbon::parse(data_get($response, 'updateTime')),
             expirationTime: Carbon::parse(data_get($response, 'expirationTime')),
